backend/requirements.txt
asgiref==3.3.1
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
cryptography==37.0.2
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
icalendar==4.0.9
idna==2.10
numpy==1.21.5
pandas==1.3.5
Pillow==8.2.0
pkg-resources==0.0.0
psycopg2==2.8.6
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.1
requests==2.25.1
scipy==1.7.3
six==1.16.0
sqlparse==0.4.1
stripe==2.64.0
urllib3==1.26.6
